Fix alerts not triggering in integration when diffing binary files.
In seechanges.c, the "files changed" alert will not be generated when
diff's return code is not 1 (when system returns 256):

    if(system(diff_cmd) != 256)
    {
        merror("%s: ERROR: Unable to run diff for %s->%s",
               ARGV0,  host, script);
        return(NULL);
    }

    return(gen_diff_alert(filename, date_of_change));

Depending on the diff(1) used, calling diff on two binary files will
return an error code of 2.

    $ diff /bin/bash /bin/sh
    Binary files /bin/bash and /bin/sh differ
    - exit 2

In this case, these alerts appear in the log files but do not trigger
our slack integration.

Fix OSSEC's handling of diffing binary files: respond to a return code
that isn't 0 or 1 by running "diff -q" on the same files. "diff -q"
should only report when files are different, and works for binary files
as well:

    $ diff -q /bin/sh /bin/bash
    Files /bin/sh and /bin/bash differ
    - exit 1

Also add checks for diff (or diff -q) terminating abnormally.

Fix also applied to similar logic in agentlessd.c.